Move views under app
