+++ app/views/_partials/events.ejs
