+++ app/views/_partials/quotes.ejs
