+++ app/views/corporate_manager.ejs
